Add files via upload
#Smarthome Control
ESP32 DHT11 Relay Blynk 1Wire
This is my college microprocessor engineering major assignment.
This is a smart home system that uses sensors  to monitor and control different aspects of a home. I use kit ESP32 as main processor, DHT11 as sensor for measuring temperature and humidity, blynk as the display block and relay as load switching block

## Demo
See a demo of this project on YouTube.

## Hardware
- ESP32
- DHT11 temperature and humidity sensor
- Blynk app
- Relay 4 channel 5v
- LED (green)


## Usage
The smart home system has the following features:

- UserState control: Allows the user to turn on or off the lights when they are at home or away from home.
- Automatically control: Capable of controlling home electrical appliances such as fans, light bulbs, air conditioners, ... automatically based on parameters of humidity, temperature

To use the system, connect all the components as described in the Wiring section and upload the code to the Arduino board. The system will automatically monitor the sensors and control the various components based on the conditions specified in the code.